% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdtime.R
\name{birdtime}
\alias{birdtime}
\title{bird time}
\usage{
birdtime(df, confidence = 0, bird.names, bw = 0.75)
}
\arguments{
\item{df}{the dataframe created by the read_birdnet_file or read_birdnet_folder function}

\item{confidence}{the minimum confidence level for the bird call identifications}

\item{bird.names}{one to six bird names, to view comparative prevalence over time/seasons}
}
\value{
a plot showing calls over time for each species input
}
\description{
creates a streamgraph showing number of calls over time for each species
}
\examples{
\dontrun{ birdtime(df, 0.5, c("Galah", "Brown Songlark", "Little Corella")) }
}
