% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{parse_birdnet_filename_datetime}
\alias{parse_birdnet_filename_datetime}
\title{Extract a start datetime from a BirdNET selection table filename}
\usage{
parse_birdnet_filename_datetime(file_name)
}
\arguments{
\item{file_name}{The full name or path of the BirdNET selection file.}
}
\value{
A POSIXct object representing the start datetime.
}
\description{
This function assumes a filename of the form:
SOMETHING_YYYYMMDD_HHMMSS.BirdNET.selection.table.txt
}
\examples{
parse_birdnet_filename_datetime("1STSMM2_20241105_050000.BirdNET.selection.table.txt")
}
