% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_confidence.R
\name{plot_confidence}
\alias{plot_confidence}
\title{Plot Confidence}
\usage{
plot_confidence(df, confidence = 0, bw = 0.75)
}
\arguments{
\item{df}{The dataframe created by the read_birdnet_file or
read_birdnet_folder function.}

\item{confidence}{Optional - the minimum confidence level for the bird call
identifications.}

\item{bw}{Bandwidth for the density estimation (default: 0.75).}
}
\value{
A streamgraph indicating confidence of calls over time.
}
\description{
Creates a streamgraph showing confidence of recorded calls over time.
}
\examples{
\dontrun{
plot_confidence(df, confidence = 0)
}
}
