% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_counts.R
\name{plot_species_counts}
\alias{plot_species_counts}
\title{Plot Species Counts}
\usage{
plot_species_counts(
  df,
  confidence = 0,
  remove.dominants = FALSE,
  log.scale = TRUE
)
}
\arguments{
\item{df}{The dataframe created by the read_birdnet_file or
read_birdnet_folder function}

\item{confidence}{The minimum confidence level for the species identifications}

\item{remove.dominants}{Removes dominant species from plot for cleaner
visualisation (deprecated if log.scale is TRUE)}

\item{log.scale}{Logical. If TRUE (default), uses a log10 scale for the y-axis,
which handles dominant species visualization automatically.}
}
\value{
A plot showing the number of calls per species + prints removed
species if applicable
}
\description{
Takes BirdNET data and creates a figure representing number of recordings of
each species, over the given confidence interval.
}
\examples{
\dontrun{
plot_species_counts(df, 0.5)
}

}
