% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickstats.R
\name{quickstats}
\alias{quickstats}
\title{quick statistics}
\usage{
quickstats(df, confidence = 0)
}
\arguments{
\item{df}{the dataframe created by the read_birdnet_file or read_birdnet_folder function}

\item{confidence}{the minimum confidence level for the bird call identifications}
}
\value{
a tibble of quick statistics
}
\description{
takes birdnet .txt data and offers a few quick statistics
}
\examples{
\dontrun{ quickstats(df, 0.5) }

}
