% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicktime.R
\name{quicktime}
\alias{quicktime}
\title{quick time}
\usage{
quicktime(df, confidence = 0)
}
\arguments{
\item{df}{the dataframe created by the read_birdnet_file or read_birdnet_folder function}

\item{confidence}{the minimum confidence level for the bird call identifications}
}
\value{
a figure showing number of calls over time for the given confidence interval
}
\description{
creates a simple line graph showing number of calls over time
}
\examples{
\dontrun{ quicktime(df, 0.5) }
}
