% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise_detections}
\alias{summarise_detections}
\title{Summarise Detections}
\usage{
summarise_detections(df, confidence = 0)
}
\arguments{
\item{df}{The dataframe created by the read_birdnet_file or read_birdnet_folder function.}

\item{confidence}{The minimum confidence level for the bird call identifications.}
}
\value{
A tibble of summary statistics.
}
\description{
Takes BirdNET results and offers a few quick statistics.
}
\examples{
\dontrun{
summarise_detections(df, 0.5)
}

}
