% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stream.R
\name{plot_species_stream}
\alias{plot_species_stream}
\title{Plot Species Stream}
\usage{
plot_species_stream(
  df,
  confidence = 0,
  bird.names = unique(df$`Common Name`),
  bw = 0.75
)
}
\arguments{
\item{df}{The dataframe created by the \code{read_birdnet_file} or \code{read_birdnet_folder} function.}

\item{confidence}{The minimum confidence level for the species identifications.}

\item{bird.names}{One to six species names, to view comparative prevalence over time/seasons.}

\item{bw}{The bandwidth for the kernel density estimation performed by the streamgraph.}
}
\value{
A plot showing calls over time for each species input.
}
\description{
Creates a streamgraph showing number of calls over time for each species.
}
\examples{
\dontrun{
plot_species_stream(df, 0.5, c("Galah", "Brown Songlark", "Little Corella"))
}
}
