% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_and_combine.R
\name{read_birdnet_file}
\alias{read_birdnet_file}
\title{Read a single BirdNET selection table file}
\usage{
read_birdnet_file(file_path, tz = "UTC")
}
\arguments{
\item{file_path}{Path to a BirdNET .selection.table.txt file}

\item{tz}{Timezone if you want to override or specify a different timezone
than what is parsed from the filename (if any). Default "UTC".}
}
\value{
A tibble with all columns from the BirdNET file plus:
\itemize{
\item \code{file_name} for reference
\item \code{start_time} the parsed date-time from the filename
\item \code{recording_window_time} the absolute time of each detection
}
}
\description{
Read a single BirdNET selection table file
}
\examples{
\dontrun{
df <- read_birdnet_file("1STSMM2_20241105_050000.BirdNET.selection.table.txt")
head(df)
}
}
