% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_birdnet_folder}
\alias{read_birdnet_folder}
\title{Read all BirdNET selection files in a folder}
\usage{
read_birdnet_folder(
  folder = ".",
  pattern = "BirdNET.*\\\\.(txt|csv)$",
  recursive = FALSE
)
}
\arguments{
\item{folder}{A folder path containing BirdNET files.}

\item{pattern}{A regex pattern to match the files.
Default matches \code{.txt} or \code{.csv} files that look like BirdNET outputs.
e.g. "BirdNET.*\\.(txt|csv)$"}

\item{recursive}{Whether to search recursively in subfolders. Default FALSE.}
}
\value{
A single tibble combining all data.
}
\description{
Read all BirdNET selection files in a folder
}
\examples{
\dontrun{
all_detections <- read_birdnet_folder("path/to/folder")
}
}
