% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_and_combine.R
\name{read_birdnet_sites}
\alias{read_birdnet_sites}
\title{Read BirdNET selection files from multiple sites (folders)}
\usage{
read_birdnet_sites(
  folder_paths,
  pattern = "BirdNET.selection.table.txt$",
  recursive = FALSE
)
}
\arguments{
\item{folder_paths}{A character vector of folder paths, one for each site.}

\item{pattern}{A regex pattern to match the files. Default "BirdNET.selection.table.txt$".}

\item{recursive}{Whether to search recursively in subfolders. Default FALSE.}
}
\value{
A single tibble combining all data.
Includes a \code{Site} column derived from the folder name.
}
\description{
Read BirdNET selection files from multiple sites (folders)
}
\examples{
\dontrun{
folders <- c("detections_SiteA", "detections_SiteB")
all_sites <- read_birdnet_sites(folders, pattern = "BirdNET.results.csv$")
}
}
